refactor(lstar): split containers.py into a containers package
The monolithic containers.py (~1015 lines) held every lstar container
type in one file with no logical grouping. Split it into a containers/
package of small modules organized by domain role:

- interval, identifiers, participation: scalar units and registry sets
- aggregation: post-quantum signature aggregation proofs
- validator: genesis configuration and the validator registry
- checkpoint: Casper-FFG checkpoints and the attestation vote
- attestation, block: vote envelopes and the blocks that carry them
- state, store: per-block consensus state and the fork-choice view

The package __init__ re-exports the full public surface, so every
existing import path keeps working unchanged.

Also:
- Rename the Config container to GenesisConfig (and the protocol hook
  config_class to genesis_config_class) to disambiguate it from the
  config.py constants module. The SSZ fixture wire name stays "Config".
- Move VALIDATOR_REGISTRY_LIMIT into config.py beside HISTORICAL_ROOTS_LIMIT.
- Move the container unit tests into a mirrored tests/.../containers/
  package and update the CLAUDE.md mirroring example to the new path.
