+++ src/lean_spec/spec/forks/__init__.py
@@ -13,7 +13,7 @@ from lean_spec.spec.forks.lstar.containers import (
     BlockBody,
     BlockHeader,
     Checkpoint,
-    Config,
+    GenesisConfig,
     Interval,
     SignedAggregatedAttestation,
     SignedAttestation,
@@ -50,11 +50,11 @@ __all__ = [
     "BlockBody",
     "BlockHeader",
     "Checkpoint",
-    "Config",
     "DEFAULT_REGISTRY",
     "FORK_SEQUENCE",
     "ForkProtocol",
     "ForkRegistry",
+    "GenesisConfig",
     "IMMEDIATE_JUSTIFICATION_WINDOW",
     "Interval",
     "LstarSpec",
