+++ src/lean_spec/spec/forks/lstar/containers/state.py
@@ -1,0 +1,217 @@
+"""Consensus state and the justification/finalization accounting it tracks."""
+
+from typing import Self
+
+from lean_spec.spec.forks.lstar.config import HISTORICAL_ROOTS_LIMIT, VALIDATOR_REGISTRY_LIMIT
+from lean_spec.spec.forks.lstar.containers.block import BlockHeader
+from lean_spec.spec.forks.lstar.containers.checkpoint import Checkpoint
+from lean_spec.spec.forks.lstar.containers.validator import GenesisConfig, Validators
+from lean_spec.spec.forks.lstar.slot import Slot
+from lean_spec.spec.ssz import Boolean, Bytes32, Container, SSZList
+from lean_spec.spec.ssz.bitfields import BaseBitlist
+
+
+class HistoricalBlockHashes(SSZList[Bytes32]):
+    """List of historical block root hashes up to historical roots limit."""
+
+    LIMIT = int(HISTORICAL_ROOTS_LIMIT)
+
+
+class JustificationRoots(SSZList[Bytes32]):
+    """List of justified block roots up to historical roots limit."""
+
+    LIMIT = int(HISTORICAL_ROOTS_LIMIT)
+
+
+class JustifiedSlots(BaseBitlist):
+    """Bitlist tracking justified slots up to historical roots limit."""
+
+    LIMIT = int(HISTORICAL_ROOTS_LIMIT)
+
+    def is_slot_justified(self, finalized_slot: Slot, target_slot: Slot) -> Boolean:
+        """
+        Determine if a specific slot is considered justified.
+
+        The check follows these rules:
+        - Slots at or before the finalized boundary are implicitly justified.
+        - Future slots are checked against the tracked bitfield.
+
+        Args:
+            finalized_slot: The anchor point for the tracking window.
+            target_slot: The slot to query.
+
+        Returns:
+            True if the slot is justified or finalized, False otherwise.
+
+        Raises:
+            IndexError: If the target slot is active but outside the tracked range.
+        """
+        # First, determine the position of the target relative to the anchor.
+        #
+        # If the result is None, the slot is behind the finalized boundary.
+        # By definition, finalized slots are justified.
+        if (relative_index := target_slot.justified_index_after(finalized_slot)) is None:
+            return Boolean(True)
+
+        # Check the tracked bitfield for the slot's status.
+        #
+        # We assume the slot is within the tracked range.
+        #
+        # If the caller asks for a slot too far in the future, it indicates a logic error.
+        try:
+            return self[relative_index]
+        except IndexError as e:
+            raise IndexError(
+                f"Slot {target_slot} is outside the tracked range "
+                f"(finalized_boundary={finalized_slot}, tracked_length={len(self)})"
+            ) from e
+
+    def with_justified(
+        self,
+        finalized_slot: Slot,
+        target_slot: Slot,
+        value: Boolean,
+    ) -> Self:
+        """
+        Return a new bitfield with the justification status updated.
+
+        This method follows the immutable pattern:
+        - Returns 'self' if the slot is finalized (immutable).
+        - Returns a clone with the specific bit updated for active slots.
+
+        Args:
+            finalized_slot: The anchor point for the tracking window.
+            target_slot: The slot to update.
+            value: The new justification status.
+
+        Returns:
+            A new, updated JustifiedSlots instance.
+
+        Raises:
+            IndexError: If the target slot is active but outside the tracked range.
+        """
+        # Determine the position of the target relative to the anchor.
+        #
+        # If the slot is behind the finalized boundary, we return 'self' unchanged.
+        # We cannot modify the status of finalized history, and treating it as a
+        # no-op preserves the immutability of the conceptual chain history.
+        if (relative_index := target_slot.justified_index_after(finalized_slot)) is None:
+            return self
+
+        # Ensure we are not trying to write to a future slot that does not exist
+        # in our tracking list yet. The state must be explicitly extended first.
+        if relative_index >= len(self):
+            raise IndexError(
+                f"Slot {target_slot} is outside the tracked range "
+                f"(finalized_boundary={finalized_slot}, tracked_length={len(self)})"
+            )
+
+        # Clone and update in one smooth operation.
+        #
+        # 1. Create a shallow copy of the data list to avoid mutating the original.
+        # 2. Update the specific bit in the copy.
+        # 3. Construct a new instance with the updated data.
+        new_data = list(self.data)
+        new_data[relative_index] = value
+
+        return type(self)(data=new_data)
+
+    def extend_to_slot(self, finalized_slot: Slot, target_slot: Slot) -> Self:
+        """
+        Extend the tracking capacity to cover a new target slot.
+
+        This prepares the state to process a new block by ensuring the
+        bitfield is long enough to store its justification status.
+        Gaps are filled with False (unjustified).
+
+        Args:
+            finalized_slot: The anchor point for the tracking window.
+            target_slot: The slot that must be addressable.
+
+        Returns:
+            A new instance with sufficient capacity.
+        """
+        # Calculate the index required to store the status of the target.
+        #
+        # If the target is already finalized, no extension is needed because
+        # we don't track finalized data.
+        if (relative_index := target_slot.justified_index_after(finalized_slot)) is None:
+            return self
+
+        # Calculate how many new entries we need to append.
+        #
+        # Since indices are zero-based, the required capacity is index + 1.
+        # If we already have enough capacity, the gap will be zero or negative.
+        required_capacity = relative_index + 1
+        if (gap_size := required_capacity - len(self)) <= 0:
+            return self
+
+        # Return a new instance with the extended data list.
+        #
+        # We extend the existing data with False values to bridge the gap.
+        return type(self)(data=list(self.data) + [Boolean(False)] * gap_size)
+
+    def shift_window(self, delta: int) -> Self:
+        """
+        Advance the tracking window by dropping slots that became finalized.
+
+        A non-positive delta keeps the tracking window unchanged.
+        """
+        # If the boundary hasn't moved forward, the window stays the same.
+        if delta <= 0:
+            return self
+
+        # Return a new instance containing only the relevant subset of data.
+        return type(self)(data=self.data[delta:])
+
+
+class JustificationValidators(BaseBitlist):
+    """Per-root validator vote bitfields, concatenated into one flat bitlist.
+
+    Each tracked root contributes one bit per registered validator.
+    The cap is the maximum tracked roots times the validator registry limit.
+
+    Why this product: a larger cap inflates the bitlist's merkle tree depth.
+    That changes the state root for identical data, so this limit is consensus-critical.
+    """
+
+    LIMIT = int(HISTORICAL_ROOTS_LIMIT) * int(VALIDATOR_REGISTRY_LIMIT)
+
+
+class State(Container):
+    """The main consensus state object."""
+
+    # Configuration
+    config: GenesisConfig
+    """The chain's configuration parameters."""
+
+    # Slot and block tracking
+    slot: Slot
+    """The current slot number."""
+
+    latest_block_header: BlockHeader
+    """The header of the most recent block."""
+
+    # Checkpoints
+    latest_justified: Checkpoint
+    """The latest justified checkpoint."""
+
+    latest_finalized: Checkpoint
+    """The latest finalized checkpoint."""
+
+    # Historical data
+    historical_block_hashes: HistoricalBlockHashes
+    """A list of historical block root hashes."""
+
+    justified_slots: JustifiedSlots
+    """A bitfield indicating which historical slots were justified."""
+
+    validators: Validators
+    """Registry of validators tracked by the state."""
+
+    # Justification tracking (flattened for SSZ compatibility)
+    justifications_roots: JustificationRoots
+    """Roots of justified blocks."""
+
+    justifications_validators: JustificationValidators
+    """A bitlist of validators who participated in justifications."""
