+++ tests/lean_spec/spec/forks/lstar/containers/test_checkpoint.py
