+++ tests/lean_spec/spec/forks/lstar/containers/test_participation.py
